// src/engine/client/types.ts
type Brand<B> = { [__brand]: B };
type Branded<T, B> = T & Brand<B>;

export enum Player {
  A = "A",
  B = "B",
}

export enum CardClass {
  POKEMON = "pokemon",
  ITEM = "item",
  SUPPORTER = "supporter",
  TOOL = "tool",
}

export enum EnergyType {
  GRASS = "grass",
  FIRE = "fire",
  WATER = "water",
  LIGHTNING = "lightning",
  FIGHTING = "fighting",
  PSYCHIC = "psychic",
  DARK = "dark",
  METAL = "metal",
}

export enum PokemonType {
  GRASS = "grass",
  FIRE = "fire",
  WATER = "water",
  LIGHTNING = "lightning",
  FIGHTING = "fighting",
  PSYCHIC = "psychic",
  DARK = "dark",
  METAL = "metal",
  COLORLESS = "colorless",
  DRAGON = "dragon",
}

export type CardGameId = Branded<string, "CardGameId">;
export type CardStableId = Branded<string, "CardStableId">;

export type CardReference<T extends CardClass> = {
  cardId: CardGameId;
  cardStableId: CardStableId;
  cardClass: T;
};

export type EnergyCount = number;
export type HealthPoints = number;

export type PokemonState = {
  cardReference: CardReference<CardClass.POKEMON>;
  currentHealthPoints: HealthPoints;
  // currentStatusCondition
  // currentEffects
  attachedEnergy: Partial<Record<EnergyType, EnergyCount>>;
  attachedTool: CardReference<CardClass.TOOL> | null;
  // allowedAttacks
  evolvedFrom: CardReference<CardClass.POKEMON>[];
};

export type TurnAllowances = {
  canAttachFreeEnergyFromZone: boolean;
  canUseSupporterCard: boolean;
  canRetreat: boolean;
};

export type PrizePoints = number;

type BenchSlot = PokemonState | null;

type Bench = [BenchSlot, BenchSlot, BenchSlot];

export type InternalGameState = {
  turnNumber: TurnNumber;
  activePlayer: Player;
  currentTurnAllowances: TurnAllowances;
  prizePoints: {
    [Player.A]: PrizePoints;
    [Player.B]: PrizePoints;
  };
  active: {
    [Player.A]: PokemonState | null;
    [Player.B]: PokemonState | null;
  };
  bench: {
    [Player.A]: Bench;
    [Player.B]: Bench;
  };
  hand: {
    [Player.A]: CardReference<CardClass>[];
    [Player.B]: CardReference<CardClass>[];
  };
  discardPile: {
    [Player.A]: CardReference<CardClass>[];
    [Player.B]: CardReference<CardClass>[];
  };
  deck: {
    [Player.A]: CardReference<CardClass>[];
    [Player.B]: CardReference<CardClass>[];
  };
};

export type TurnNumber = number;
// ...
export class InvalidGameStateError extends Error {}

export type AttackParams = {
  targetCardId?: CardGameId;
};

export type AttackPreview = {
  expectedDamageToOpponentActivePokemon: number;
};

export type AttackId = Branded<string, "AttackId">;

export type AttackEnergyCost = {
  typedEnergy: Partial<Record<EnergyType, EnergyCount>>;
  colorlessEnergy: EnergyCount;
};

export type AttackConfig = {
  id: AttackId;
  name: string;
  energyCost: AttackEnergyCost;
  description: string | null;
  damageDescriptor: string;
  onUse: (game: InternalGameState, params: AttackParams) => InternalGameState;
  onPreview: (game: InternalGameState) => AttackPreview;
};